Update README.md
The educational system has had serious hit from examination malpractice.
In addition, in these days a student suspended from a school because of his misbehavior easily joins another institution with his/her parent’s support. This practice should not be encouraged. This has brought us to focus the need to tackle this issues in order to find ways to eradicate such problems in the light of the above, the study looked at examination malpractices in college  and schools with the view to achieve malpractice free examination

To determine the causes of examination malpractices in secondary schools and tertiary institutions .
To investigate on the solutions that are to be undertaken to prevent malpractice
To find an optimal solution to the above stated problem
To develop an user friendly website and make it more efficient in preventing malpractice